Added UpdateRequired to UPDATE_ON_BOOT

+++ scripts/start.sh
@@ -33,8 +33,12 @@ fi
 
 # Always update on boot even if the server is installed, to prevent appmanifest issues
 if [ "$ServerInstalled" == 0 ] && [ "${UPDATE_ON_BOOT,,}" == true ]; then
-    rm /palworld/steamapps/appmanifest_2394010.acf
-    InstallServer
+    UpdateRequired
+    updateRequired=$?
+    if [ "$updateRequired" == 0 ]; then
+        rm /palworld/steamapps/appmanifest_2394010.acf
+        InstallServer
+    fi
 fi
 
 STARTCOMMAND=("./PalServer.sh")
